float rating

+++ src/models/Rating.js
@@ -3,7 +3,7 @@ module.exports = (sequelize) => {
    // defino el modelo
    sequelize.define('rating',
       {
-         rating: { type: DataTypes.INTEGER, allowNull: false },
+         rating: { type: DataTypes.FLOAT, allowNull: false },
          emailUser: { type: DataTypes.STRING, allowNull: false },
          comment: { type: DataTypes.TEXT },
          date: { type: DataTypes.DATE, allowNull: false }
